backend/requirements.txt
flask
mysql-connector-python
markdown
passlib
uWSGI
flask-login
flask-restful
